1. changed code structure 2. finished login and register functionality 3. fix some bugs
