+++ requirements.txt
@@ -1,1 +1,3 @@
 mysql-connector-python
+PyJWT
+flask-cors
