+++ scripts/config.py
@@ -1,0 +1,3 @@
+import os
+
+SECRET_KEY = os.getenv('SECRET_KEY', '123456')
